requirements.txt
pandas
plotly
strands-agents
strands-agents-tools
mcp
